requirements.txt
cudatoolkit=10.2
pytorch=1.5
numpy
matplotlib
argparse
tsort